Front-andBackend/praexiois/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
drf-yasg==1.21.6
idna==3.4
inflection==0.5.1
ldap3==2.9.1
packaging==23.1
pyasn1==0.5.0
PyJWT==2.7.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==1.26.15
